src/mcp_server/requirements.txt
python-dotenv
gql
mcp
aiohttp